requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==20.1.0
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
flask-ngrok==0.0.25
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gspread==3.7.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.1
numpy==1.20.0
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.2.1
pdfkit==0.6.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
PyMySQL==0.10.0
pyngrok==5.0.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
SQLAlchemy==1.3.17
twilio==6.52.0
urllib3==1.25.10
Werkzeug==0.16.1
zipp==3.1.0
